Can now add a shot! Dataframe and qt model are updated with shot data. Server works but dataframe appears to be malformed for the conversion that happens on the other side. This is likely because the shot is one that has not run yet and so does not have many layers to the multiindex. This should probably be considered a bug though - it should handle it regardless.
Might move the processing server side if that's possible.

methods are written in a re-usable way such that they can be used by the rest
of the program as it is ported over.

Timezone is now looked up instead of hardcoded (removed google's tz module)
Float values are rendered in scientific notation