Fixes #17 by speeding up dataframe transfers.
Moved call to df.convert_objects to the lyse server, which makes dataframes more
efficiently picklable by replacing columns containing only a single array datatype
with an array of that datatype. Previously df.convert_obects was only called
after the transfer, so the preceding pickling was slow.

For the large, mostly numeric dataframe tested, this speeds
up pickling and unpickling by approx a factor of 5.

Also set convert_numeric=False in the call to df.convert_objects. It was
previously set to True, but I don't think this is what we want. With it False,
all-numeric columns are still converted as desired. With it True, *strings* are
converted to numbers even if only a single string is convertible to a numeric
value, with all non-convertable values in the column becoming NaNs. I don't
think this is what we want. If the user asked for a string, the data should
remain a string. It is presumptuous to assume otherwise and I suspect a
misunderstanding by whoever wrote this line originally.