FIxed bugs in and made more robust the copying to clipboard functionality. The figure is saved to disk and then a subprocess is used to copy it to the clipboard, and the subprocess keeps running until the clipboard is overwritten, so that pasting works on setups which require the program to provide the paste data at paste time (and thus require them to keep running). Cough Xorg, cough.
This might sound like overkill, but It's much cleaner. The subprocess can have
its own QApplication to handle the clipboard actions, and thus doesn't
interfere with the parent's QApplication, and doesn't require it to have one,
if matplotlib is using a different backend (even the Qt backend could be qt 4 or
qt 5 at the moment, so we don't want to have to deal with this). I found
difficult to debug breakage when trying to make this work when a QApplication
already existed, so a clean break is much nicer, and is required to keep the
clipboard data around (on linux anyway) once the copying program has closed.

I've made the subprocess code forward compatible with PyQt5.

This functionality requires a function I added to zprocess 2.2, so I added
a check_version call for this.