Removed debugging printlines.
Reverted to use exec(code, snadbox) directly instead of six.exec_, since the
tuple form of the exec statement is valid Python 2 that is forward compatible
with exec being a function in Python 3.

Modified the traceback logic to not hit on a CPython bug in CPython < 2.7.9
that made Python throw an error about the exec.