Fix bug where restoring the window position/size when a different number of monotors or different monitor resolution is in use compared to when that information was saved could potentially move the window off the screen.
Made it so that if the number of screens and their sizes is not the same
as when the data was saved, to simply fall back on defaults.

This way those with a consistent setup will get lyse in the same position
they left it, and those with a varying setup will get the defaults after
their setup changes.