Support integer indexing of the lyse DataFrame, using a MultiIndex of: ['sequence_index', 'run number', 'run repeat']
This can be enabled in labconfig using:
   [lyse]
   integer_indexing = true

If this section or option are absent from the labconfig file, fallback to the
existing indexing by ['sequence', 'run time'].

This feature supports much easier use of pd.DataFrame.loc, e.g.
   df = lyse.data()
   df.loc[56]       # sequence 56 MultiIndexed by ['run number', 'repeat number']
   df.loc[56:60]    # sequences 56 through 60 inclusive
   df.loc[56,4]     # run number 4 of sequence 56, indexed by 'repeat number'
   df.loc[56,2:6,:] # runs 2 through 6 (inclusive) of sequence 56

Enabling this feature changes the vertical header item to read, e.g.
   105. 0056 | 0001 | 0106

This is far more useful, as these integers can all be used to construct a .loc
or .iloc call to the lyse DataFrame.

The main problem with this feature is that the MultiIndex might not be unique if
new shots are sent to or loaded into lyse with an existing sequence_index, e.g.
if the sequence_index gets reset in runmanager or shots are loaded from multiple
days.