Added the ability for analysis scripts to register a custom Plot class
This allows you to provide custom behaviour in plot windows (such as additional actions)

Use `lyse.register_plot_class(identifier, cls)` to register a class for lyse to use when creating window for `figure(identifier)`, where cls must be a subclass of `lyse.Plot` (which is monkey patched in to the lyse module to avoid the double import denier script triggering if you try to import it directly from lyse.analysis_subprocess)

Currently we do not support changing the class from one run of the script to the next, without restarting the worker process (as that would require re-instantiating the plot which has a large number of problems). Also, in order to avoid the custom subclass being detected as changing from shot to shot, it needs to be imported from a separate module to the analysis script. If an invalid class is specified, then the default `Plot` class is used and message is printed in the lyse terminal.

I also changed the indexing of the internal storage of plots to use the identifier rather than the fig object. This seems to make much more sense to me, and since we also now expose that dictionary to the analysis script via `lyse.plots`, I figured now would be a good time to change it! I can't see any downside to doing this but please correct me if I'm wrong.

Finally, this obviously has no effect on lyse scripts run from the terminal, since we don't use our figure manager in such cases. A warning is printed informing the user in such instances (if they use `lyse.register_plot_class()`)