Disable word wrap in FileBox QTableView
Line wrapping in the FileBox is not desired since we are not expanding cells
vertically, and if line wrapping is enabled the text is vertically centered
and so not even the first line is readable.

QTableView has word wrapping enabled by default, but Qt <= 5.11 had a bug such
that it had no effect. Qt 5.12 fixes this bug, resulting in the table wrapping
lines unless word wrapping is explicitly disabled as in this commit.