An attempt to fix issue #45.
Although I don't see why the Qt model or dataframe might sometimes not update,
I noticed that the calls to update_row with updated_row_data provided were
*not* actually updating the Qt model, because the check for whether a column
was in the updated_row_data dictionary was comparing a column tuple padded with
empty strings to one not padded with empty strings. Further, the call to update
the progress percentage was updating *all* qt model items for the row.

Now, the two things are fixed - the updated_row_data does actually update
the qt model items, and the update of the status percentage does not.

There is no reason to believe this will fix issue #45 other than it touching
code that could be relevant and was not working as intended. I still don't
see specifically what could cause issue #45.

Also, the updated data dict sent back to lyse from analysis routines should
only include items under the 'results/' group in the HDF5 file. So this fixes
that too.