Ensure number of levels in the dataframe does not decrease.
This fixes a bug where the number of levels is incorrect after all shots are
deleted and shots with fewer levels are loaded. We could allow the number of
levels to decrease, but this is more trouble than it is worth since we store
previously-seen columns so that their visibility can be restored if they are
seen again. Since their levels cannot be reduced, comparing them with other
columns is difficult. There is surely a cleaner way to deal with all this,
but it will require more work, whereas the bug is fixed by simply ensuring that
the number of column levels never shrinks.